+++ tools/storage/Cargo.toml
@@ -26,7 +26,6 @@ diem-push-metrics = { workspace = true }
 diem-types = { workspace = true }
 glob = { workspace = true }
 libra-backwards-compatibility = { workspace = true }
-libra-types = { workspace = true }
 num_cpus = { workspace = true }
 serde_json = { workspace = true }
 tokio = { workspace = true }
