+++ warehouse/Cargo.toml
@@ -27,11 +27,11 @@ log = { workspace = true }
 once_cell = { workspace =  true }
 serde = { workspace =  true }
 serde_json = { workspace =  true }
-sqlx = { workspace = true }
+
 tokio = { workspace = true }
 neo4rs = { workspace = true }
-cypher-dto = "0.3.1"
 
 
 [dev-dependencies]
 testcontainers = { workspace = true }
+sqlx = { workspace = true }
