+++ warehouse/tests/experimental/mod.rs
@@ -1,0 +1,10 @@
+#[allow(dead_code)]
+pub mod load_account;
+#[allow(dead_code)]
+pub mod load_coin;
+#[allow(dead_code)]
+pub mod pg_migrate;
+#[allow(dead_code)]
+pub mod query_balance;
+#[allow(dead_code)]
+pub mod restaurant;
