+++ warehouse/tests/experimental/query_balance.rs
@@ -2,7 +2,7 @@ use anyhow::Result;
 use libra_types::exports::AccountAddress;
 use sqlx::PgPool;
 
-use crate::table_structs::WarehouseBalance;
+use libra_warehouse::table_structs::WarehouseBalance;
 
 // TODO: return specific commit errors for this batch
 pub async fn query_last_balance(
@@ -21,7 +21,12 @@ pub async fn query_last_balance(
         "#
     );
 
-    let row: WarehouseBalance = sqlx::query_as(&query_template).fetch_one(pool).await?;
+    let row = sqlx::query(&query_template).fetch_one(pool).await?;
+    dbg!(&row);
 
-    Ok(row)
+    let dummy = WarehouseBalance {
+        balance: 0,
+    };
+
+    Ok(dummy)
 }
