+++ warehouse/tests/support/mod.rs
@@ -4,3 +4,4 @@ pub mod fixtures;
 pub mod pg_testcontainer;
 
 pub mod neo4j_testcontainer;
+pub mod age_init;
